requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
bitarray==3.4.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
hexbytes==1.3.0
idna==3.10
inference-sdk==0.49.1
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.4.3
mypy_extensions==1.1.0
mysql-connector==2.2.9
mysql-connector-python==9.3.0
namex==0.0.9
numpy==2.1.3
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
parsimonious==0.10.0
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.4
py-cpuinfo==9.0.0
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rlp==4.1.0
scipy==1.15.3
setuptools==80.3.1
simple-websocket==1.1.0
six==1.17.0
supervision==0.25.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uuid==1.30
web3==7.11.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
